Add Spotify music dataset
This commit adds the dataset file dataset.csv containing Spotify music data to the repository. The dataset will be used for data analysis and machine learning model training during the project. It was sourced from Kaggle and will be kept in the /data directory of the repository for easy access and reference.